Star-Office-UI/skills/prism-alerts/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
SQLAlchemy>=2.0.0
cryptography>=41.0.0
